[Refactor] grid/feature-action-flags.cpp/h をsystem/terrain/terrain-actions-table.cpp/h に改名した
