+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -2187,9 +2187,6 @@
     <ClCompile Include="..\..\src\system\grid-type-definition.cpp">
       <Filter>system</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\grid\feature-action-flags.cpp">
-      <Filter>grid</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\avatar\avatar-changer.cpp">
       <Filter>avatar</Filter>
     </ClCompile>
@@ -2549,6 +2546,9 @@
     </ClCompile>
     <ClCompile Include="..\..\src\system\floor\town-list.cpp">
       <Filter>system\floor</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\system\terrain\terrain-actions-table.cpp">
+      <Filter>system\terrain</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -5013,9 +5013,6 @@
     <ClInclude Include="..\..\src\system\grid-type-definition.h">
       <Filter>system</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\grid\feature-action-flags.h">
-      <Filter>grid</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\monster\monster-damage.h">
       <Filter>monster</Filter>
     </ClInclude>
@@ -5523,6 +5520,9 @@
     <ClInclude Include="..\..\src\game-option\game-option-page.h">
       <Filter>game-option</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\system\terrain\terrain-actions-table.h">
+      <Filter>system\terrain</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\src\wall.bmp" />
@@ -5813,6 +5813,9 @@
     </Filter>
     <Filter Include="system\enums\terrain">
       <UniqueIdentifier>{3ceaf1e1-713b-4cfe-979d-e0ecd9949bfc}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="system\terrain">
+      <UniqueIdentifier>{3a15d903-e2ff-488d-a3c8-7bf69fad0f43}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
