+++ src/grid/grid.cpp
@@ -27,7 +27,6 @@
 #include "game-option/game-play-options.h"
 #include "game-option/map-screen-options.h"
 #include "game-option/special-options.h"
-#include "grid/feature-action-flags.h"
 #include "grid/feature.h"
 #include "grid/object-placer.h"
 #include "grid/trap.h"
@@ -51,6 +50,7 @@
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "system/terrain-type-definition.h"
+#include "system/terrain/terrain-actions-table.h"
 #include "term/gameterm.h"
 #include "term/term-color-types.h"
 #include "timed-effect/timed-effects.h"
